+++ src/paco/models/schemas.py
@@ -6593,6 +6593,39 @@ CodeCommit DeploymentPipeline Source Stage
         required=False,
     )
 
+class IDeploymentPipelineSourceGitHub(IDeploymentPipelineStageAction):
+    """GitHub DeploymentPipeline Source Stage
+
+To configure a GitHub source, first create an OAuth Token in your GitHub account with the scopes
+``repo`` and ``admin:repoHook``. See the AWS documentation on how to `Configure Your Pipeline to Use a Personal Access Token`_.
+
+Using the AWS Console for the account and region that the CodePipeline will be provisioned in, go to AWS Systems Manager
+and in Parameter Store choose "create parameter" and create a SecureString. Store the OAuth Token created by GitHub as the value
+of this parameter. Put the name you gave your paramter in the ``github_token_parameter_name`` field.
+
+.. _Configure Your Pipeline to Use a Personal Access Token: https://docs.aws.amazon.com/codepipeline/latest/userguide/GitHub-create-personal-token-CLI.html
+
+"""
+    taggedValue('contains', 'mixed')
+    deployment_branch_name = schema.TextLine(
+        title="The name of the branch where source changes are to be detected.",
+        description="",
+        default="master",
+        required=False,
+    )
+    github_owner = schema.TextLine(
+        title='The name of the GitHub user or organization who owns the GitHub repository.',
+        required=True,
+    )
+    github_repository = schema.TextLine(
+        title='The name of the repository where source changes are to be detected.',
+        required=True,
+    )
+    github_token_parameter_name = schema.TextLine(
+        title='The name of the SSM Parameter SecureString that contains the GitHub OAuth Token.',
+        required=True,
+    )
+
 class IDeploymentPipelineBuildCodeBuild(IDeploymentPipelineStageAction):
     """
 CodeBuild DeploymentPipeline Build Stage